// src/types/api.d.ts
export interface paths {
// ...
  "/api/v1/user-coupon/my": {
    /** getMyCoupons */
    get: operations["getMyCouponsUsingGET"];
  };
// ...
}

export type webhooks = Record<string, never>;

export interface components {
  schemas: {
    /** ArticleAlarmRegisterDto */
    ArticleAlarmRegisterDto: {
      articleUuid?: string;
      /** @enum {string} */
      entityStatus?: "ACTIVE" | "BLOCKED" | "DELETED" | "INVALID" | "SEARCHABLE";
    };
    /** ArticleContentFileDto */
    ArticleContentFileDto: {
      fileName?: string;
      fileS3Key?: string;
      fileUrl?: string;
    };
    /** ArticleEditDto */
    ArticleEditDto: {
      changeLog?: string;
      content?: string;
// ...
    };
// ...
    MentorViewDto: {
      availableSlots?: components["schemas"]["MentorAvailableSlotDto"][];
      createdBy?: string;
      /** Format: date-time */
      createdDate?: string;
      description?: string;
      educations?: components["schemas"]["MentorEducationDto"][];
      elapsedCreatedDate?: string;
      /** @enum {string} */
      entityStatus?: "ACTIVE" | "BLOCKED" | "DELETED" | "INVALID" | "SEARCHABLE";
      experiences?: components["schemas"]["MentorExperienceDto"][];
      hashTags?: string[];
      isBookmarkedByMe?: boolean;
      lastModifiedBy?: string;
      /** Format: date-time */
      lastModifiedDate?: string;
      /** Format: int32 */
      maxDiscountedAblePrice?: number;
      name?: string;
      /** Format: int32 */
      originPrice?: number;
      profileImageUrl?: string;
      recurringSlots?: components["schemas"]["MentorRecurringSlotViewDto"][];
      shortDescription?: string;
      uuid?: string;
    };
// ...
    UserCouponViewDto: {
      coupon?: components["schemas"]["CouponViewDto"];
      createdBy?: string;
      /** Format: date-time */
      createdDate?: string;
      elapsedCreatedDate?: string;
      /** @enum {string} */
      entityStatus?: "ACTIVE" | "BLOCKED" | "DELETED" | "INVALID" | "SEARCHABLE";
      lastModifiedBy?: string;
      /** Format: date-time */
      lastModifiedDate?: string;
      uuid?: string;
    };
// ...
    UserViewDto: {
      birth?: string;
      categoryRefType?: string;
      createdBy?: string;
      /** Format: date-time */
      createdDate?: string;
      description?: string;
      elapsedCreatedDate?: string;
      email?: string;
      emailToChange?: string;
      /** @enum {string} */
      entityStatus?: "ACTIVE" | "BLOCKED" | "DELETED" | "INVALID" | "SEARCHABLE";
      files?: components["schemas"]["FAEntityFileViewDto"][];
      /** Format: int64 */
      followId?: number;
      followedByMe?: boolean;
      /** @enum {string} */
      gender?: "FEMALE" | "MALE" | "NONE";
      interestCategories?: ("BUSINESS" | "POLITICS" | "SPORT")[];
      isAdmin?: boolean;
      isBlocked?: boolean;
      isFollowedByMe?: boolean;
      isPushNotificationOn?: boolean;
      isSecret?: boolean;
      lastModifiedBy?: string;
      /** Format: date-time */
      lastModifiedDate?: string;
      /** Format: int64 */
      likeCount?: number;
      major?: string;
      mentorUuid?: string;
      name?: string;
      /** @enum {string} */
      paymentStatus?: "NONE" | "PLAN_A";
      phone?: string;
      profileImage?: string;
      profileImageUrl?: string;
      /** @enum {string} */
      provider?: "APP" | "APPLE" | "FACEBOOK" | "KAKAOTALK" | "TOSS";
      realName?: string;
      /** Format: date-time */
      recentFollowDate?: string;
      /** Format: int64 */
      reviewCount?: number;
      school?: components["schemas"]["SchoolTypeViewDto"];
      schoolCertificateFiles?: components["schemas"]["FAEntityFileViewDto"][];
      schoolName?: string;
      schoolType?: string;
      /** @enum {string} */
      schoolVerificationStatus?: "APPROVED" | "IN_REVIEW" | "NONE" | "REJECTED";
      /** @enum {string} */
      status?: "ACTIVATED" | "DEACTIVATED" | "NOT_VERIFIED";
      /** @enum {string} */
      studentType?: "MASTER" | "NONE" | "PHD" | "POSTDOCTOR" | "POSTGRAD" | "PROFESSOR" | "UNDERGRADUATE";
      unregisteredReason?: string;
      /** Format: int32 */
      userAvailablePrice?: number;
      userReports?: components["schemas"]["UserReportViewDto"][];
      uuid?: string;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}

export type $defs = Record<string, never>;

export type external = Record<string, never>;

export interface operations {

  /** registerArticleAlarm */
  registerArticleAlarmUsingPOST: {
    requestBody?: {
      content: {
        "application/json": components["schemas"]["ArticleAlarmRegisterDto"];
      };
    };
    responses: {
      /** @description OK */
      200: {
        content: {
          "*/*": boolean;
        };
      };
      /** @description Created */
      201: {
        content: never;
      };
      /** @description Unauthorized */
      401: {
        content: never;
      };
      /** @description Forbidden */
      403: {
        content: never;
      };
      /** @description Not Found */
      404: {
        content: never;
      };
    };
  };
// ...
}